Add existing string format validations
